refactor(ZNTA-2116): Moved TriCheckboxGroup to own file
