+++ server/zanata-frontend/src/frontend/app/components/TriCheckbox/TriCheckbox.story.js
@@ -2,8 +2,8 @@ import React, {Component} from 'react'
 import PropTypes from 'prop-types'
 import { storiesOf, action } from '@kadira/storybook'
 import RealTriCheckbox from '.'
-import { Table, ListGroup, ListGroupItem } from 'react-bootstrap'
-import { find } from 'lodash'
+import TriCheckboxGroup from './TriCheckboxGroup'
+import { Table } from 'react-bootstrap'
 import './index.css'
 
 class TriCheckbox extends Component {
@@ -66,111 +66,6 @@ class WithButton extends Component {
           onChange={this.checkboxChanged}
         />
         <button onClick={this.makeIndeterminate}>Make intermediate</button>
-      </div>
-    )
-  }
-}
-
-class TriCheckboxGroup extends Component {
-  static propTypes = {
-    checked: PropTypes.bool.isRequired,
-    indeterminate: PropTypes.bool.isRequired,
-    options: PropTypes.arrayOf(PropTypes.shape({
-      checked: PropTypes.bool,
-      name: PropTypes.string
-    })),
-    onClick: PropTypes.func.isRequired
-  }
-  static defaultProps = {
-    checked: false,
-    indeterminate: false,
-    options: [{checked: false, name: 'nothing'}],
-    onClick: () => {
-    }
-  }
-
-  constructor (props) {
-    super(props)
-    this.state = {
-      checked: this.props.checked,
-      indeterminate: this.props.indeterminate,
-      options: this.props.options
-    }
-  }
-
-  masterCheckBoxClick = (event) => {
-    event.persist()
-    if (event.target.checked) {
-      this.setState((prevState, props) => ({
-        checked: true,
-        indeterminate: false,
-        options: props.options.map((value) => {
-          value.checked = true
-          return value
-        })
-      }))
-    } else {
-      this.setState((prevState, props) => ({
-        checked: false,
-        indeterminate: false,
-        options: props.options.map((value) => {
-          value.checked = false
-          return value
-        })
-      }))
-    }
-    this.props.onClick(event)
-  }
-
-  subCheckboxClick = (event) => {
-    event.persist()
-    this.setState((prevState, props) => ({
-      options: props.options.map((value) => {
-        if (event.target.name === value.name) {
-          value.checked = !value.checked
-        }
-        return value
-      }),
-      checked: props.options.every((option) => {
-        return option.checked
-      }),
-      indeterminate: !(props.options.every((option) => {
-        return option.checked
-      })) && (find(props.options, (option) => {
-        return option.checked
-      }) !== undefined)
-    }))
-    this.props.onClick(event)
-  }
-
-  render () {
-    const optionGroup = this.props.options.map((value, index) => {
-      return (
-        <ListGroupItem key={index}>
-          <RealTriCheckbox
-            name={value.name}
-            checked={this.state.options[index].checked}
-            indeterminate={false}
-            onChange={this.subCheckboxClick}
-          />
-          {value.name}
-        </ListGroupItem>
-      )
-    })
-    return (
-      <div>
-        <h3>Checkbox group with tri-state control</h3>
-        <ListGroup>
-          <ListGroupItem>
-            <RealTriCheckbox
-              checked={this.state.checked}
-              indeterminate={this.state.indeterminate}
-              onChange={this.masterCheckBoxClick}
-            />
-            <label>Fruits</label>
-          </ListGroupItem>
-          {optionGroup}
-        </ListGroup>
       </div>
     )
   }
